feat: allow file append

+++ internal/templates/writer.go
@@ -40,36 +40,69 @@ func (f *FileWriter) Write(ctx context.Context, ui *slog.Logger, templatedFiles 
 	egrp, _ := errgroup.WithContext(ctx)
 	for _, file := range templatedFiles {
 		egrp.Go(func() error {
-			if _, err := os.Stat(file.DestinationPath); err != nil {
-				if !errors.Is(err, os.ErrNotExist) {
-					return fmt.Errorf("failed to check if file exists: %s, %w", file.DestinationPath, err)
-				}
-			} else {
-				if f.promptOverride != nil {
-					fileExistsLock.Lock()
-					defer fileExistsLock.Unlock()
+			switch file.Mode {
+			case TemplatedFileWriteModeFile:
+				if _, err := os.Stat(file.DestinationPath); err != nil {
+					if !errors.Is(err, os.ErrNotExist) {
+						return fmt.Errorf("failed to check if file exists: %s, %w", file.DestinationPath, err)
+					}
+				} else {
+					if f.promptOverride != nil {
+						fileExistsLock.Lock()
+						defer fileExistsLock.Unlock()
 
-					override, err := f.promptOverride(file)
+						override, err := f.promptOverride(file)
+						if err != nil {
+							return fmt.Errorf("failed to get answer to whether a file should be overwritten or not: %w", err)
+						}
+
+						if !override {
+							ui.Warn("Skipping file", "file", file.DestinationPath)
+							return nil
+						}
+					}
+				}
+
+				if parent := path.Dir(file.DestinationPath); parent != "" && parent != "/" {
+					if err := os.MkdirAll(parent, readWriteExec); err != nil {
+						return fmt.Errorf("failed to create parent dir for: %s, %w", file.DestinationPath, err)
+					}
+				}
+
+				ui.Info("writing file", "path", file.DestinationPath)
+				if err := os.WriteFile(file.DestinationPath, file.Content, readExec); err != nil {
+					return fmt.Errorf("failed to write file: %s, %w", file.DestinationPath, err)
+				}
+			case TemplatedFileWriteModeAppend:
+				_, err := os.Stat(file.DestinationPath)
+				if err != nil {
+					if !errors.Is(err, os.ErrNotExist) {
+						return fmt.Errorf("failed to check if file exists: %s, %w", file.DestinationPath, err)
+					}
+				}
+
+				if errors.Is(err, os.ErrNotExist) {
+					if parent := path.Dir(file.DestinationPath); parent != "" && parent != "/" {
+						if err := os.MkdirAll(parent, readWriteExec); err != nil {
+							return fmt.Errorf("failed to create parent dir for: %s, %w", file.DestinationPath, err)
+						}
+					}
+
+					ui.Info("writing file", "path", file.DestinationPath)
+					if err := os.WriteFile(file.DestinationPath, file.Content, readExec); err != nil {
+						return fmt.Errorf("failed to write file: %s, %w", file.DestinationPath, err)
+					}
+				} else {
+					ui.Info("appending file", "path", file.DestinationPath)
+					fileDest, err := os.OpenFile(file.DestinationPath, os.O_APPEND|os.O_CREATE|os.O_WRONLY, readExec)
 					if err != nil {
-						return fmt.Errorf("failed to get answer to whether a file should be overwritten or not: %w", err)
+						return fmt.Errorf("open file for append %s: %w", file.DestinationPath, err)
 					}
 
-					if !override {
-						ui.Warn("Skipping file", "file", file.DestinationPath)
-						return nil
+					if _, err := fileDest.Write(file.Content); err != nil {
+						return fmt.Errorf("append to file %s: %w", file.DestinationPath, err)
 					}
 				}
-			}
-
-			if parent := path.Dir(file.DestinationPath); parent != "" && parent != "/" {
-				if err := os.MkdirAll(parent, readWriteExec); err != nil {
-					return fmt.Errorf("failed to create parent dir for: %s, %w", file.DestinationPath, err)
-				}
-			}
-
-			ui.Info("writing file", "path", file.DestinationPath)
-			if err := os.WriteFile(file.DestinationPath, file.Content, readExec); err != nil {
-				return fmt.Errorf("failed to write file: %s, %w", file.DestinationPath, err)
 			}
 
 			return nil
